finally fix all tests
